+++ lib/src/export.dart
@@ -80,6 +80,7 @@ export './sankey/sankey_chart.dart';
 export './sunburst/sunburst_chart.dart';
 export './distribution/box_plot_chart.dart';
 export './distribution/histogram_chart.dart';
+export './bubble/bubble_chart.dart';
 export './gauge/gauge_chart.dart';
 export './treemap/treemap_chart.dart';
 export './treemap/treemap_data.dart';